fix: incremental_vacuum 漏 fetchall 致 SQLite 空闲页不回收 + 解码去除多余 astype 拷贝
PRAGMA incremental_vacuum 靠结果集逐页驱动回收,execute 不 fetchall 只 step
一次、每次仅回收 1 页。main.py 两处(observability.db / miloco.db 的 cleanup)
均漏 fetchall,空闲页持续堆积——实测 observability.db 空闲 63%、miloco.db 24%。
两处补 .fetchall()。

decoder.py: to_ndarray(format="bgr24") 已返回独立 uint8 数组,再 .astype("uint8")
默认 copy=True 会对每帧多做一次全量拷贝,去除。

+++ backend/miot/src/miot/decoder.py
@@ -292,7 +292,9 @@ class MIoTMediaDecoder(threading.Thread):
         if self._video_frame_callback and frames:
             for frame in frames:
                 try:
-                    bgr = frame.to_ndarray(format="bgr24").astype("uint8")
+                    # to_ndarray(bgr24) 已返回独立的 uint8 数组，再 .astype("uint8")
+                    # 默认 copy=True 会对每帧多做一次全量拷贝，纯浪费。
+                    bgr = frame.to_ndarray(format="bgr24")
                 except Exception as e:
                     _LOGGER.warning("Failed to convert frame to ndarray: %s", e)
                     continue
